+++ UCR.Core/Context.cs
@@ -33,6 +33,7 @@ namespace HidWizards.UCR.Core
 
         public delegate void ActiveProfileChanged(Profile profile);
         public event ActiveProfileChanged ActiveProfileChangedEvent;
+        public event Action MinimizedToTrayEvent;
         
         internal bool IsNotSaved { get; private set; }
         internal IOController IOController { get; set; }
@@ -68,7 +69,8 @@ namespace HidWizards.UCR.Core
         private void SetCommandLineOptions()
         {
             options = new OptionSet {
-                { "p|profile=", "The profile to search for", FindAndLoadProfile }
+                { "p|profile=", "The profile to search for", FindAndLoadProfile },
+                { "h|hidden", "Minimize to system tray", x => MinimizedToTrayEvent.Invoke() }
             };
         }