+++ UCR/App.xaml.cs
@@ -19,6 +19,7 @@ namespace HidWizards.UCR
         private Context context;
         private HidGuardianClient _hidGuardianClient;
         private SingleGlobalInstance mutex;
+        private bool StartMinimized;
 
         protected override void OnStartup(StartupEventArgs e)
         {
@@ -35,15 +36,21 @@ namespace HidWizards.UCR
                 InitializeUcr();
                 CheckForBlockedDll();
 
-                context.ParseCommandLineArguments(e.Args);
                 var mw = new MainWindow(context);
-                mw.Show();
+                context.MinimizedToTrayEvent += Context_MinimizedToTrayEvent;
+                context.ParseCommandLineArguments(e.Args);
+                if (!StartMinimized) mw.Show();
             }
             else
             {
                 SendArgs(string.Join(";", e.Args));
                 Current.Shutdown();
             }
+        }
+
+        private void Context_MinimizedToTrayEvent()
+        {
+            StartMinimized = true;
         }
 
         private void InitializeUcr()